address Jenny comments
