+++ micro_dl/cli/inference_script.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env/python
 """Model inference on larger images with and w/o stitching"""
 import argparse
+import glob
+import os
 import yaml
 
 from micro_dl.inference import image_inference as image_inf
@@ -30,7 +32,7 @@ def parse_args():
         help='data split to predict on'
     )
     parser.add_argument(
-        '--inf_config',
+        '--config',
         type=str,
         help='path to inference yaml configuration file',
     )
@@ -48,13 +50,23 @@ def run_prediction(args, gpu_ids, gpu_mem_frac):
     :return:
     """
 
-    config_fname = args.inf_config
+    config_fname = args.config
     with open(config_fname, 'r') as f:
-        inf_config = yaml.load(f)
+        inf_config = yaml.safe_load(f)
 
+    train_config_fname = glob.glob(
+        os.path.join(inf_config['model_dir'], '*.yml')
+    )
+    assert len(train_config_fname) == 1, \
+        'more than one train config yaml found in model dir'
+
+    model_fname = None
+    if model_fname in inf_config:
+        model_fname = os.path.join(inf_config['model_dir'],
+                                   inf_config['model_fname'])
     image_pred_inst = image_inf.ImagePredictor(
-        config=inf_config['train_config'],
-        model_fname=inf_config['model_fname'],
+        config=train_config_fname[0],
+        model_fname=model_fname,
         image_dir=inf_config['image_dir'],
         data_split=args.data_split,
         image_param_dict=inf_config['image_params_dict'],
