+++ micro_dl/deprecated/create_segmentation_masks.py
